:shirt: Refactor view namespaces
