+++ src/SceneGate.UI.Formats/Binary/DesignTextViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace SceneGate.UI.Formats.Common;
+﻿namespace SceneGate.UI.Formats.Binary;
 
 using System;
 using Yarhl.IO;
