+++ src/SceneGate.UI.Formats/Binary/HexViewerView.axaml
@@ -2,9 +2,9 @@
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-             xmlns:local="using:SceneGate.UI.Formats.Common"
+             xmlns:local="using:SceneGate.UI.Formats.Binary"
              mc:Ignorable="d" d:DesignWidth="1050" d:DesignHeight="600"
-             x:Class="SceneGate.UI.Formats.Common.HexViewerView"
+             x:Class="SceneGate.UI.Formats.Binary.HexViewerView"
              x:DataType="local:HexViewerViewModel">
   <Design.DataContext>
     <local:HexViewerViewModel />
