+++ src/SceneGate.UI.Formats/Binary/HexViewerViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace SceneGate.UI.Formats.Common;
+﻿namespace SceneGate.UI.Formats.Binary;
 
 using System;
 using System.Buffers.Binary;
