+++ src/SceneGate.UI.Formats/CommonFormatsViewModelBuilder.cs
@@ -4,7 +4,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using SceneGate.UI.Formats.Binary;
-using SceneGate.UI.Formats.Common;
 using SceneGate.UI.Formats.Graphics;
 using SceneGate.UI.Formats.Texts;
 using Texim.Compressions.Nitro;
