+++ src/SceneGate.UI.Formats/Models/ObjectView.axaml.cs
@@ -1,4 +1,4 @@
-﻿namespace SceneGate.UI.Formats.Common;
+﻿namespace SceneGate.UI.Formats.Models;
 
 using Avalonia.Controls;
