+++ src/SceneGate.UI/Pages/Main/AnalyzeViewModel.cs
@@ -12,7 +12,8 @@ using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using SceneGate.UI.ControlsData;
 using SceneGate.UI.Formats;
-using SceneGate.UI.Formats.Common;
+using SceneGate.UI.Formats.Binary;
+using SceneGate.UI.Formats.Models;
 using SceneGate.UI.Formats.Graphics;
 using SceneGate.UI.Formats.Mvvm;
 using SceneGate.UI.Pages.Analyze;
